+++ index.php
@@ -15,42 +15,42 @@ $ads = [
         'category' => 'Доски и лыжи',
         'price' => 10999,
         'img' => 'img/lot-1.jpg',
-        'date_expiration' => '2020-07-18'
+        'date_expiration' => '2020-07-21'
     ],
     [
         'title' => 'DC Ply Mens 2016/2017 Snowboard',
         'category' => 'Доски и лыжи',
         'price' => 159999,
         'img' => 'img/lot-2.jpg',
-        'date_expiration' => '2020-07-19'
+        'date_expiration' => '2020-07-21'
     ],
     [
         'title' => 'Крепления Union Contact Pro 2015 года размер L/XL',
         'category' => 'Крепления',
         'price' => 8000,
         'img' => 'img/lot-3.jpg',
-        'date_expiration' => '2020-07-20'
+        'date_expiration' => '2020-07-22'
     ],
     [
         'title' => 'Ботинки для сноуборда DC Mutiny Charocal',
         'category' => 'Ботинки',
         'price' => 10999,
         'img' => 'img/lot-4.jpg',
-        'date_expiration' => '2020-07-21'
+        'date_expiration' => '2020-07-23'
     ],
     [
         'title' => 'Куртка для сноуборда DC Mutiny Charocal',
         'category' => 'Одежда',
         'price' => 7500,
         'img' => 'img/lot-5.jpg',
-        'date_expiration' => '2020-07-22'
+        'date_expiration' => '2020-07-24'
     ],
     [
         'title' => 'Маска Oakley Canopy',
         'category' => 'Разное',
         'price' => 5400,
         'img' => 'img/lot-6.jpg',
-        'date_expiration' => '2020-07-23'
+        'date_expiration' => '2020-07-25'
     ]
 ];
