+++ templates/main.php
@@ -1,16 +1,5 @@
 <?php
-function calculate_time_to_expiration ($date_expiration) {
-    $date_expiration = strtotime($date_expiration);
-    $current_date = strtotime(date('Y-m-d h:i:s'));
-    $diff = $date_expiration - $current_date;
-    $hours_expiration = floor($diff / 3600);
-    if ($hours_expiration < 10) {
-        $hours_expiration = str_pad($hours_expiration, 2, "0", STR_PAD_LEFT);
-    };
-    $minutes_expiration = floor(($diff - ($hours_expiration * 3600)) / 60);
-
-    return [$hours_expiration, $minutes_expiration];
-}
+require_once('functions/functions.php');
 ?>
 
 <section class="promo">
@@ -42,8 +31,11 @@ function calculate_time_to_expiration ($date_expiration) {
                             <span class="lot__amount">Стартовая цена</span>
                             <span class="lot__cost"><?=format_price($val['price']); ?></span>
                         </div>
-                        <div class="lot__timer timer <?php if (calculate_time_to_expiration($val['date_expiration'])[0] == 0): ?>timer--finishing<?php endif ?>">
-                            <?=calculate_time_to_expiration($val['date_expiration'])[0] . ':' . calculate_time_to_expiration($val['date_expiration'])[1]?>
+                        <?php
+                          $time_to_expiration = calculate_time_to_expiration($val['date_expiration']);
+                        ?>
+                        <div class="lot__timer timer <?php if ($time_to_expiration[0] == 0): ?>timer--finishing<?php endif ?>">
+                            <?=$time_to_expiration[0] . ':' . $time_to_expiration[1]?>
                         </div>
                     </div>
                 </div>
